+++ srcs/ft_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: tcherret <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/11 16:54:50 by tcherret          #+#    #+#             */
-/*   Updated: 2018/12/11 17:07:01 by tcherret         ###   ########.fr       */
+/*   Updated: 2018/12/14 19:09:58 by tcherret         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putnbr_base(unsigned long long nb, int base)
 {
-	if (nb < base)
+	if (nb < (unsigned long long)base)
 	{
 		if (nb < 10)
 			ft_putchar('0' + nb);
