+++ srcs/ft_putnbr_basemaj.c
@@ -6,7 +6,7 @@
 /*   By: tcherret <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/11 17:07:08 by tcherret          #+#    #+#             */
-/*   Updated: 2018/12/11 17:35:31 by tcherret         ###   ########.fr       */
+/*   Updated: 2018/12/14 19:10:27 by tcherret         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putnbr_basemaj(unsigned long long nb, int base)
 {
-	if (nb < base)
+	if (nb < (unsigned long long)base)
 	{
 		if (nb < 10)
 			ft_putchar('0' + nb);
